can now update quanitity css time...

+++ pantry-tracker/app/page.js
@@ -35,7 +35,7 @@ export default function Home() {
     const docs = await getDocs(snapshot);
     const pantryList = [];
     docs.forEach(doc => {
-      pantryList.push(doc.id);
+      pantryList.push({name: doc.id, ...doc.data()});
     });
     console.log(pantryList);
     setPantry(pantryList);
@@ -47,14 +47,31 @@ export default function Home() {
 
   const addItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await setDoc(docRef, {})
-    updatePantry()
+
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      await setDoc(docRef, {count: count + 1})
+    }
+    else{
+      await setDoc(docRef, {count: 1})
+    }
+    await updatePantry()
   };
 
   const removeItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await deleteDoc(docRef)
-    updatePantry()
+
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if (count === 1){
+        await deleteDoc(docRef)
+      } else{
+        await setDoc(docRef, {count: count - 1})
+      }
+    await updatePantry()
+    }
   }
 
   return (
@@ -107,9 +124,9 @@ export default function Home() {
         </Box>
 
         <Stack width = "800px" height = "300px" spacing= {2} overflow= {'auto'}>
-          {pantry.map((i) => (
+          {pantry.map(({name, count}) => (
               <Box
-                key = {i}
+                key = {name}
                 bgcolor = {'#f0f0f0'}
                 color = {'black'}
                 display = {'flex'}
@@ -122,11 +139,15 @@ export default function Home() {
                 <Typography variant= {'h3'} color = {'#333'} textAlign = {'center'}>
                   {
                     // capitalize the first letter of the item
-                    i.charAt(0).toUpperCase() + i.slice(1)
+                    name.charAt(0).toUpperCase() + name.slice(1)
                   }
                 </Typography>
 
-                <Button variant="contained" onClick={() => removeItem(i)}>
+                <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+                  Quantity: {count}
+                </Typography>
+
+                <Button variant="contained" onClick={() => removeItem(name)}>
                   Remove
                 </Button>
               </Box>
